requirements.txt
bcrypt==3.1.1
blinker==1.4
cffi==1.8.3
click==6.6
coverage==4.2
factory-boy==2.7.0
fake-factory==0.7.2
Faker==0.7.3
Flask==0.11.1
Flask-DebugToolbar==0.10.0
Flask-SQLAlchemy==2.1
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
passlib==1.6.5
pkg-resources==0.0.0
psycopg2==2.6.2
pycparser==2.17
python-dateutil==2.5.3
six==1.10.0
SQLAlchemy==1.1.3
Werkzeug==0.11.11
